refactor(jest-runtime): extract data: URI and import-attribute code from EsmLoader
Both blocks are self-contained: the WHATWG data URL decoding moves to
dataUri.ts, and import-attribute validation moves to importAttributes.ts
as ImportAttributeValidator. The validator's warned-pairs set was
module-scoped, deduplicating the JSON-without-attribute deprecation
warning across every Runtime in a worker - the second and later test
files never saw it. Instance state on the per-Runtime validator restores
the warning per test file.
