+++ packages/jest-runtime/src/internals/__tests__/EsmLoader.test.ts
@@ -15,7 +15,7 @@ import {
 import type {JestEnvironment} from '@jest/environment';
 import {invariant} from 'jest-util';
 import type {CjsExportsCache} from '../CjsExportsCache';
-import {EsmLoader, LOAD_ASYNC, validateImportAttributes} from '../EsmLoader';
+import {EsmLoader, LOAD_ASYNC} from '../EsmLoader';
 import {CjsParseError} from '../ModuleExecutor';
 import type {FileCache} from '../FileCache';
 import type {JestGlobals} from '../JestGlobals';
@@ -726,230 +726,40 @@ describe('EsmLoader.dynamicImportFromCjs (legacy linkAndEvaluate)', () => {
   );
 });
 
-describe('validateImportAttributes', () => {
-  // Each test uses a unique modulePath/referencingIdentifier pair so the
-  // deprecation-warn dedup cache (module-scoped in EsmLoader) doesn't suppress
-  // across tests.
-  let counter = 0;
-  const uniquePaths = () => {
-    counter += 1;
-    return {
-      js: `/test-${counter}.js`,
-      json: `/test-${counter}.json`,
-      referencer: `/referencer-${counter}.mjs`,
-    };
-  };
-
-  describe('JSON modules', () => {
-    test('accepts type: json', () => {
-      const {json, referencer} = uniquePaths();
-      expect(() =>
-        validateImportAttributes(json, {type: 'json'}, referencer),
-      ).not.toThrow();
-    });
-
-    test('throws ERR_IMPORT_ATTRIBUTE_TYPE_INCOMPATIBLE when type is wrong', () => {
-      const {json, referencer} = uniquePaths();
-      let error: NodeJS.ErrnoException | null = null;
-      try {
-        validateImportAttributes(json, {type: 'css'}, referencer);
-      } catch (error_) {
-        error = error_ as NodeJS.ErrnoException;
-      }
-      expect(error).not.toBeNull();
-      expect(error).toBeInstanceOf(TypeError);
-      expect(error?.code).toBe('ERR_IMPORT_ATTRIBUTE_TYPE_INCOMPATIBLE');
-      expect(error?.message).toMatch(/not of type "css"/);
-    });
-
-    test('warns once per (referencer, module) when no attribute is present', () => {
-      const {json, referencer} = uniquePaths();
-      const warnSpy = jest.spyOn(console, 'warn').mockImplementation(() => {});
-      try {
-        validateImportAttributes(json, {}, referencer);
-        validateImportAttributes(json, {}, referencer);
-        validateImportAttributes(json, {}, referencer);
-        expect(warnSpy).toHaveBeenCalledTimes(1);
-        expect(warnSpy).toHaveBeenCalledWith(
-          expect.stringContaining(
-            'importing JSON without an import attribute is deprecated',
-          ),
-        );
-      } finally {
-        warnSpy.mockRestore();
-      }
-    });
-
-    test('warns again for a different referencer importing the same module', () => {
-      const {json, referencer} = uniquePaths();
-      const otherReferencer = `${referencer}-other`;
-      const warnSpy = jest.spyOn(console, 'warn').mockImplementation(() => {});
-      try {
-        validateImportAttributes(json, {}, referencer);
-        validateImportAttributes(json, {}, otherReferencer);
-        expect(warnSpy).toHaveBeenCalledTimes(2);
-      } finally {
-        warnSpy.mockRestore();
-      }
-    });
-
-    test('treats data:application/json URIs as JSON modules', () => {
-      const {referencer} = uniquePaths();
-      const dataUri = 'data:application/json,{"x":1}';
-      // type: 'json' is accepted
-      expect(() =>
-        validateImportAttributes(dataUri, {type: 'json'}, referencer),
-      ).not.toThrow();
-      // Wrong type is rejected
-      let error: NodeJS.ErrnoException | null = null;
-      try {
-        validateImportAttributes(dataUri, {type: 'css'}, referencer);
-      } catch (error_) {
-        error = error_ as NodeJS.ErrnoException;
-      }
-      expect(error?.code).toBe('ERR_IMPORT_ATTRIBUTE_TYPE_INCOMPATIBLE');
-    });
-
-    test('truncates data: URI payload in the deprecation warning', () => {
-      const {referencer} = uniquePaths();
-      const huge = 'a'.repeat(10_000);
-      const dataUri = `data:application/json,${encodeURIComponent(huge)}`;
-      const warnSpy = jest.spyOn(console, 'warn').mockImplementation(() => {});
-      try {
-        validateImportAttributes(dataUri, {}, referencer);
-        expect(warnSpy).toHaveBeenCalledTimes(1);
-        const message = warnSpy.mock.calls[0][0];
-        expect(message).toContain('data:application/json,…');
-        expect(message).not.toContain(huge);
-      } finally {
-        warnSpy.mockRestore();
-      }
-    });
-
-    test('warning mentions both static and dynamic syntax', () => {
-      const {json, referencer} = uniquePaths();
-      const warnSpy = jest.spyOn(console, 'warn').mockImplementation(() => {});
-      try {
-        validateImportAttributes(json, {}, referencer);
-        const message = warnSpy.mock.calls[0][0];
-        expect(message).toContain("with { type: 'json' }");
-        expect(message).toContain("{ with: { type: 'json' } }");
-      } finally {
-        warnSpy.mockRestore();
-      }
-    });
-  });
-
-  describe('non-JSON modules', () => {
-    test('accepts no attributes', () => {
-      const {js, referencer} = uniquePaths();
-      expect(() => validateImportAttributes(js, {}, referencer)).not.toThrow();
-    });
-
-    test('throws ERR_IMPORT_ATTRIBUTE_TYPE_INCOMPATIBLE when type is set', () => {
-      const {js, referencer} = uniquePaths();
-      let error: NodeJS.ErrnoException | null = null;
-      try {
-        validateImportAttributes(js, {type: 'javascript'}, referencer);
-      } catch (error_) {
-        error = error_ as NodeJS.ErrnoException;
-      }
-      expect(error?.code).toBe('ERR_IMPORT_ATTRIBUTE_TYPE_INCOMPATIBLE');
-      expect(error?.message).toMatch(/not of type "javascript"/);
-    });
-
-    test('throws when type: json is asserted on non-JSON', () => {
-      const {js, referencer} = uniquePaths();
-      let error: NodeJS.ErrnoException | null = null;
-      try {
-        validateImportAttributes(js, {type: 'json'}, referencer);
-      } catch (error_) {
-        error = error_ as NodeJS.ErrnoException;
-      }
-      expect(error?.code).toBe('ERR_IMPORT_ATTRIBUTE_TYPE_INCOMPATIBLE');
-    });
-  });
-
-  describe('unknown attribute keys', () => {
-    test('throws ERR_IMPORT_ATTRIBUTE_UNSUPPORTED on a JSON module', () => {
-      const {json, referencer} = uniquePaths();
-      let error: NodeJS.ErrnoException | null = null;
-      try {
-        validateImportAttributes(
-          json,
-          {cache: 'no-store', type: 'json'},
-          referencer,
-        );
-      } catch (error_) {
-        error = error_ as NodeJS.ErrnoException;
-      }
-      expect(error?.code).toBe('ERR_IMPORT_ATTRIBUTE_UNSUPPORTED');
-      expect(error?.message).toMatch(/Import attribute "cache"/);
-    });
-
-    test('throws ERR_IMPORT_ATTRIBUTE_UNSUPPORTED on a non-JSON module', () => {
-      const {js, referencer} = uniquePaths();
-      let error: NodeJS.ErrnoException | null = null;
-      try {
-        validateImportAttributes(js, {foo: 'bar'}, referencer);
-      } catch (error_) {
-        error = error_ as NodeJS.ErrnoException;
-      }
-      expect(error?.code).toBe('ERR_IMPORT_ATTRIBUTE_UNSUPPORTED');
-    });
-
-    test('rejects unknown key before reporting type-mismatch', () => {
-      // Unknown-key check is first per Node's `validateAttributes` order.
-      const {json, referencer} = uniquePaths();
-      let error: NodeJS.ErrnoException | null = null;
-      try {
-        validateImportAttributes(
-          json,
-          {nonsense: 'x', type: 'css'},
-          referencer,
-        );
-      } catch (error_) {
-        error = error_ as NodeJS.ErrnoException;
-      }
-      expect(error?.code).toBe('ERR_IMPORT_ATTRIBUTE_UNSUPPORTED');
-    });
-  });
-
-  describe('requireResultFromModule', () => {
-    test('hands out the raw namespace when sync evaluation is unavailable', () => {
-      const {stubs} = makeLoader();
-      jest.isolateModules(() => {
-        jest.doMock('../nodeCapabilities', () => ({
-          ...jest.requireActual<typeof import('../nodeCapabilities')>(
-            '../nodeCapabilities',
-          ),
-          supportsSyncEvaluate: false,
-        }));
-        const {EsmLoader: GatedEsmLoader} = require('../EsmLoader');
-        const loader = new GatedEsmLoader({
-          cjsExportsCache: stubs.cjsExportsCache,
-          coreModule: stubs.coreModule,
-          environment: stubs.environment,
-          fileCache: stubs.fileCache,
-          jestGlobals: stubs.jestGlobals,
-          mockState: stubs.mockState,
-          registries: stubs.registries,
-          requireModule: stubs.requireModule,
-          requireModuleOrMock: stubs.requireModuleOrMock,
-          resolution: stubs.resolution,
-          shouldLoadAsEsm: stubs.shouldLoadAsEsm,
-          testPath: '/test.js',
-          testState: stubs.testState,
-          transformCache: stubs.transformCache,
-        });
-        const namespace = {default: 'D', x: 1};
-        const module = {
-          identifier: '/m.mjs',
-          namespace,
-          status: 'evaluated',
-        };
-        expect(loader.requireResultFromModule(module)).toBe(namespace);
+describe('requireResultFromModule', () => {
+  test('hands out the raw namespace when sync evaluation is unavailable', () => {
+    const {stubs} = makeLoader();
+    jest.isolateModules(() => {
+      jest.doMock('../nodeCapabilities', () => ({
+        ...jest.requireActual<typeof import('../nodeCapabilities')>(
+          '../nodeCapabilities',
+        ),
+        supportsSyncEvaluate: false,
+      }));
+      const {EsmLoader: GatedEsmLoader} = require('../EsmLoader');
+      const loader = new GatedEsmLoader({
+        cjsExportsCache: stubs.cjsExportsCache,
+        coreModule: stubs.coreModule,
+        environment: stubs.environment,
+        fileCache: stubs.fileCache,
+        jestGlobals: stubs.jestGlobals,
+        mockState: stubs.mockState,
+        registries: stubs.registries,
+        requireModule: stubs.requireModule,
+        requireModuleOrMock: stubs.requireModuleOrMock,
+        resolution: stubs.resolution,
+        shouldLoadAsEsm: stubs.shouldLoadAsEsm,
+        testPath: '/test.js',
+        testState: stubs.testState,
+        transformCache: stubs.transformCache,
       });
+      const namespace = {default: 'D', x: 1};
+      const module = {
+        identifier: '/m.mjs',
+        namespace,
+        status: 'evaluated',
+      };
+      expect(loader.requireResultFromModule(module)).toBe(namespace);
     });
   });
 });
